more debug more tunning... still troubles!

+++ conf.c
@@ -387,7 +387,7 @@ void load_conf(struct meterec_s *meterec) {
 	unsigned int port_list_len, connection_list_len;
 	const char *takes, *record, *name, *port_name, *time;
 	int mute=OFF, thru=OFF;
-	int sample_rate;
+	long int sample_rate;
 	char fn[4];
 				
 	fprintf(meterec->fd_log,"Loading '%s'\n", meterec->conf_file);
